Critical BGG client fixes
* **FIX:** The Board Game page works again. The `bgg` dependency broke at some point due to upstream BGG API changes. As a result, it was removed and the BGG API is now hit directly in the fetcher, instead.
* **FIX:** The server temporarily becoming unreachable or unavailable would all items to disappear and an infinite loading spinner to appear on any previously loaded pages, forcing a page reload to resolve.
* **FIX:** Going “back” in the browser after clicking a board game hid all items and caused an infinite loading spinner to appear, forcing a page reload to resolve.
* **FIX:** Board games moving from the wishlist to the collection (and visa versa) now functions properly.
* **BUMP:** All npm dependencies upgraded.

+++ backend/db/CachedMongoClient.js
@@ -247,12 +247,16 @@ class CachedMongoClient {
 
     removeCollectionBoardGameById(boardGameId) {
         this.boardGameCollectionCache.delete(boardGameId);
-        return this.mongoClient.deleteById(this.propertyManager.boardGamesCollectionName, boardGameId);
+        if (!this.boardGameWishlistCache.has(boardGameId)) {
+            return this.mongoClient.deleteById(this.propertyManager.boardGamesCollectionName, boardGameId);
+        }
     }
 
     removeWishlistBoardGameById(boardGameId) {
-        this.boardGameCollectionCache.delete(boardGameId);
-        return this.mongoClient.deleteById(this.propertyManager.boardGamesCollectionName, boardGameId);
+        this.boardGameWishlistCache.delete(boardGameId);
+        if (!this.boardGameCollectionCache.has(boardGameId)) {
+            return this.mongoClient.deleteById(this.propertyManager.boardGamesCollectionName, boardGameId);
+        }
     }
 
     findBoardGames(query) {
@@ -270,9 +274,17 @@ class CachedMongoClient {
          * If the board game already exists in the cache, update it.
          */
         if (this.boardGameWishlistCache.has(upsertedBoardGame._id)) {
-            this.boardGameWishlistCache.set(upsertedBoardGame._id, upsertedBoardGame);
+            if (upsertedBoardGame.inWishlist) {
+                this.boardGameWishlistCache.set(upsertedBoardGame._id, upsertedBoardGame);
+            } else {
+                this.boardGameCollectionCache.set(upsertedBoardGame._id, upsertedBoardGame);
+            }
         } else if (this.boardGameCollectionCache.has(upsertedBoardGame._id)) {
-            this.boardGameCollectionCache.set(upsertedBoardGame._id, upsertedBoardGame);
+            if (upsertedBoardGame.inWishlist) {
+                this.boardGameWishlistCache.set(upsertedBoardGame._id, upsertedBoardGame);
+            } else {
+                this.boardGameCollectionCache.set(upsertedBoardGame._id, upsertedBoardGame);
+            }
         } else {
             if (upsertedBoardGame.inWishlist) {
                 this.boardGameWishlistCache.set(upsertedBoardGame._id, upsertedBoardGame);
